* Added safety check to ac-fs3i checking if times, steps and dts of all subfields match.
* In the large timescale now only the results of the structure-scatra field is written.

* Changed SSI convergence criterion for FS3I and FPS3I from an OR to an AND: Residual AND Increment must now be below the given tolerances.

* Some restructuring and comments.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22243 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_fs3i/fs3i_partitioned.cpp
@@ -261,7 +261,7 @@ Teuchos::ParameterList& FS3I::PartFS3I::ManipulateFsiTimeParams(const Teuchos::P
 
   if (fsisubcycles != 1) //if we have subcycling for ac_fsi
   {
-    timeparams.set<double>("TIMESTEP", fs3idyn.get<double>("TIMESTEP")/fsisubcycles);
+    timeparams.set<double>("TIMESTEP", fs3idyn.get<double>("TIMESTEP")/(double)fsisubcycles);
   }
   return timeparams;
 }
